Added template for application structure.
Added template for application structure. Moved previous logic to new
application structure.

Indomian <blade39@yandex.ru>
