+++ .gitignore
@@ -27,3 +27,6 @@
 *.exe
 *.out
 *.app
+
+#IDE
+.idea
